requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.8.0rc2
Django==5.1.5
django-celery-beat==2.8.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==25.1
django-phonenumber-field==8.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
executing==2.2.0
freezegun==1.5.5
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
ijson==3.4.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.53
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
postal-codes-tools==1.1.3
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.10.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
